Add tests for the cost calculator

+++ Shortest_Path/Algorithm/CostCalculator/BaseCostCalculator.cs
@@ -4,7 +4,7 @@ namespace Shortest_Path.Algorithm.CostCalculator
 {
     public class BaseCostCalculator : ICostCalculator
     {
-        public double GetCost(Options option, Edge cnn, Station station)
+        public decimal GetCost(Options option, Edge cnn, Station station)
         {
             return cnn.Cost;
         }
