+++ Shortest_Path/Models/Edge.cs
@@ -2,8 +2,8 @@ namespace Shortest_Path.Models
 {
     public class Edge
     {
-        public double Length { get; set; }
-        public double Cost { get; set; }
+        public int Length { get; set; }
+        public decimal Cost { get; set; }
         public Station ConnectedStation { get; set; }
     }
 }
