+++ Shortest_Path/Models/Station.cs
@@ -18,7 +18,7 @@ namespace Shortest_Path.Models
         public List<string> StationCodes { get; }
         public List<Edge> Connections { get; set; }
         public Station NearestToStart { get; set; }
-        public double? MinimumCost { get; set; }
+        public decimal? MinimumCost { get; set; }
         public bool Visited { get; set; }
 
         public Station AddStationCode(string stationCode)
